+++ jpa-weekly/src/main/java/com/example/jpaweekly/repository/CustomerRepository.java
@@ -1,7 +1,8 @@
-package com.example.jpaweekly.domain.customer.repository;
+package com.example.jpaweekly.repository;
 
-import com.example.jpaweekly.domain.customer.Customer;
+import com.example.jpaweekly.domain.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
 }
